requirements.txt
aniso8601==9.0.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==4.4.2
Flask==3.0.2
Flask-RESTful==0.3.10
html2image==2.0.4.3
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
moviepy==1.0.3
mysql-connector-python==8.3.0
numpy==1.26.4
pillow==10.3.0
proglog==0.1.10
PyMySQL==1.1.0
PythonVideoConverter==1.0.3
pytz==2024.1
requests==2.31.0
six==1.16.0
tqdm==4.66.2
urllib3==2.2.1
websocket-client==1.7.0
Werkzeug==3.0.1
zipp==3.18.1
